make cookie-based of keycloak testing possible through the app, prep for keycloak 3.4.3 (#5783)
With this change, along with some changes to the `login.turbinelabs.io` domain, it's possible to cookie-test keycloak from the app, and the app plays well with keycloak 3.4.3.

The last time I tried to do this we didn't have CORS in NGINX, and keyclaok doesn't support custom headers in CORS responses, so it just wasn't possible. Now that we have envoy and CORS support in Houston, it's possible.

The changes look like this:

- in devenv, instead of proxying login through the app proxy in `app.turbinelab.int`, actually change the hostname of the login host, so we can test with both `login.turbinelabs.int` and `login.turbinelabs.io` for `start:offline` and `start` respectively. This is necessary for local login testing to work with keycloak 3.4.3, which set cookies in the login roundtrip.
- teach the app about the X-Tbn-Keycloak-Version and X-Tbn-Keycloak-Color headers
- make AuthObj.js pass these headers when the appropriate cookies are set
- add `scope=oidc` to the API calls from the app to keycloak. These are needed in 3.4.3.
- remove support for X-Tbn-Proxy-Version, which never really worked anyway
- change the way API and Stats API headers are selected in `api.prod.js`; select them rather than deselecting everything else (now that there is more of everything else)

fbshipit-source-id: 0157455